feat:从属管理功能的完善。 包括点位管理、区域管理、合作商管理； 物理外键优化为逻辑外键； 优化前端布局。

+++ DiKeDeMS/dkd-parent/dkd-manage/src/main/java/com/dkd/manage/service/impl/PartnerServiceImpl.java
@@ -1,13 +1,18 @@
 package com.dkd.manage.service.impl;
 
 import java.util.List;
+
+import com.dkd.common.constant.ExceptionConstant;
+import com.dkd.common.exception.base.BaseException;
 import com.dkd.common.utils.DateUtils;
 import com.dkd.manage.domain.Vo.PartnerVO;
+import com.dkd.manage.mapper.NodeMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.dkd.manage.mapper.PartnerMapper;
 import com.dkd.manage.domain.Partner;
 import com.dkd.manage.service.IPartnerService;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * 合作商管理Service业务层处理
@@ -15,12 +20,16 @@ import com.dkd.manage.service.IPartnerService;
  * @author RichDu
  * @date 2024-11-28
  */
+
+@Transactional
 @Service
 public class PartnerServiceImpl implements IPartnerService 
 {
     @Autowired
     private PartnerMapper partnerMapper;
 
+    @Autowired
+    private NodeMapper nodeMapper;
     /**
      * 查询合作商管理
      * 
@@ -80,6 +89,11 @@ public class PartnerServiceImpl implements IPartnerService
     @Override
     public int deletePartnerByIds(Long[] ids)
     {
+        //判断是否有被使用的合作商
+        for (Long id : ids) {
+            if(nodeMapper.selectNodeByPartnerId(id)>0)
+                throw new BaseException(ExceptionConstant.PARTNER_HAS_NODE);
+        }
         return partnerMapper.deletePartnerByIds(ids);
     }
 
@@ -103,6 +117,9 @@ public class PartnerServiceImpl implements IPartnerService
     @Override
     public int deletePartnerById(Long id)
     {
+        //判断是否有被使用的合作商
+        if(nodeMapper.selectNodeByPartnerId(id)>0)
+            throw new BaseException(ExceptionConstant.PARTNER_HAS_NODE);
         return partnerMapper.deletePartnerById(id);
     }
 }
